refactor: internal coordinate base class
